requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.4.0
attrs==23.2.0
Babel==2.15.0
backoff==2.2.1
certifi==2024.6.2
charset-normalizer==3.3.2
courlan==1.2.0
datasets==2.19.2
dateparser==1.2.0
dill==0.3.8
distro==1.9.0
filelock==3.14.0
frozenlist==1.4.1
fsspec==2024.3.1
h11==0.14.0
haystack-ai==2.2.0
htmldate==1.8.1
httpcore==1.0.5
httpx==0.27.0
huggingface-hub==0.23.2
idna==3.7
Jinja2==3.1.4
joblib==1.4.2
jusText==3.0.1
lazy-imports==0.3.1
lxml==5.2.2
lxml_html_clean==0.1.1
MarkupSafe==2.1.5
monotonic==1.6
more-itertools==10.2.0
mpmath==1.3.0
multidict==6.0.5
multiprocess==0.70.16
networkx==3.3
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.5.40
nvidia-nvtx-cu12==12.1.105
openai==1.31.0
packaging==24.0
pandas==2.2.2
pillow==10.3.0
posthog==3.5.0
pyarrow==16.1.0
pyarrow-hotfix==0.6
pydantic==2.7.3
pydantic_core==2.18.4
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.1
regex==2024.5.15
requests==2.32.3
safetensors==0.4.3
scikit-learn==1.5.0
scipy==1.13.1
sentence-transformers==3.0.0
six==1.16.0
sniffio==1.3.1
sympy==1.12.1
tenacity==8.3.0
threadpoolctl==3.5.0
tld==0.13
tokenizers==0.19.1
torch==2.3.0
tqdm==4.66.4
trafilatura==1.10.0
transformers==4.41.2
triton==2.3.0
typing_extensions==4.12.1
tzdata==2024.1
tzlocal==5.2
urllib3==2.2.1
xxhash==3.4.1
yarl==1.9.4
aiohttp==3.9.5
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.4.0
asttokens==2.4.1
attrs==23.2.0
Babel==2.15.0
backoff==2.2.1
beautifulsoup4==4.12.3
certifi==2024.6.2
charset-normalizer==3.3.2
comm==0.2.2
courlan==1.2.0
datasets==2.19.2
dateparser==1.2.0
debugpy==1.8.1
decorator==5.1.1
dill==0.3.8
distro==1.9.0
executing==2.0.1
filelock==3.14.0
frozenlist==1.4.1
fsspec==2024.3.1
gdown==5.2.0
h11==0.14.0
haystack-ai==2.2.0
htmldate==1.8.1
httpcore==1.0.5
httpx==0.27.0
huggingface-hub==0.23.2
idna==3.7
ipykernel==6.29.4
ipython==8.25.0
jedi==0.19.1
Jinja2==3.1.4
joblib==1.4.2
jupyter_client==8.6.2
jupyter_core==5.7.2
jusText==3.0.1
lazy-imports==0.3.1
lxml==5.2.2
lxml_html_clean==0.1.1
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib-inline==0.1.7
mdit-plain==1.0.1
mdurl==0.1.2
monotonic==1.6
more-itertools==10.2.0
mpmath==1.3.0
multidict==6.0.5
multiprocess==0.70.16
nest-asyncio==1.6.0
networkx==3.3
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.5.40
nvidia-nvtx-cu12==12.1.105
openai==1.31.0
packaging==24.0
pandas==2.2.2
parso==0.8.4
pexpect==4.9.0
pillow==10.3.0
platformdirs==4.2.2
posthog==3.5.0
prompt_toolkit==3.0.46
psutil==5.9.8
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==16.1.0
pyarrow-hotfix==0.6
pydantic==2.7.3
pydantic_core==2.18.4
Pygments==2.18.0
pypdf==4.2.0
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
pyzmq==26.0.3
regex==2024.5.15
requests==2.32.3
safetensors==0.4.3
scikit-learn==1.5.0
scipy==1.13.1
sentence-transformers==3.0.0
six==1.16.0
sniffio==1.3.1
soupsieve==2.5
stack-data==0.6.3
sympy==1.12.1
tenacity==8.3.0
threadpoolctl==3.5.0
tld==0.13
tokenizers==0.19.1
torch==2.3.0
tornado==6.4
tqdm==4.66.4
trafilatura==1.10.0
traitlets==5.14.3
transformers==4.41.2
triton==2.3.0
typing_extensions==4.12.1
tzdata==2024.1
tzlocal==5.2
urllib3==2.2.1
wcwidth==0.2.13
xxhash==3.4.1
yarl==1.9.4
wikipedia
accelerate
colorama
pydantic
langdetect
ragas-haystack
deepeval